Rename : File is devided by dictionary. there is execute dictionary, pipe dictionary
