+++ execute/utile.c
@@ -1,0 +1,36 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utile.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yeoshin <yeoshin@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/26 18:57:50 by yeoshin           #+#    #+#             */
+/*   Updated: 2024/03/27 02:15:45 by yeoshin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "execute.h"
+
+void	free_list(t_list *head)
+{
+	t_list	*temp;
+
+	while (head != NULL)
+	{
+		temp = head;
+		head = head->next;
+		free(temp->content);
+		free(temp);
+	}
+}
+
+void	*ft_malloc(int size)
+{
+	void	*content;
+
+	content = malloc(size);
+	if (content == NULL)
+		exit (1);
+	return (content);
+}
